[BUGFIX release] Update to Handlebars 1.1.2.
Since this only affects the tooling, we don't need to republish a 1.2.1
for this, but I tagged it with `[BUGFIX release]` so that we can get the
stable branch `.gemspec` updated. Once merged, I can publish an updated
gem as 1.2.0.1.

Also, uses allows future point releases to be allowed for consumer
projects (ember-rails, ember-appkit-rails, etc).